requirements.txt
langchain-openai
langchain-core
python-dotenv
streamlit
langchain-community
flask
flask_session